LayoutTests/fast/dom/MicroData/001-expected.txt

This test ensures that document.getItems DOM API must exist.

PASS NodeList is defined.
PASS successfullyParsed is true

TEST COMPLETE
